requirements.txt
requests==2.28.2
pygithub==1.56
python-dotenv==0.21.0